fix: mudando taxa de regra de imposto de int para decimal

+++ BatizadoDoNovato/Entities/RegraImposto.cs
@@ -3,6 +3,6 @@ public class RegraImposto
 {
     public int Codigo { get; set; }
     public string Nome { get; set; }
-    public int Taxa { get; set; }
+    public decimal Taxa { get; set; }
     public virtual IList<ProdutoRegraImposto>? ProdutosRegrasImpostos { get; set; }
 }
